requirements.txt
 pandas
 numpy
 tqdm
 sentence-transformers
 torch
 torchvision
 torchaudio
 streamlit
 scikit-learn
 pillow
 requests
 kaggle
 joblib